--- app.py
@@ -1,100 +1,0 @@
-from cloudipsp import Api, Checkout
-from flask import Flask, render_template, request, redirect, flash, url_for
-from flask_sqlalchemy import SQLAlchemy
-from flask_migrate import Migrate
-from config import Config
-from forms.forms import LoginForm
-
-app = Flask(__name__)
-app.config.from_object(Config)
-db = SQLAlchemy(app)
-migrate = Migrate(app, db)
-# from app import db
-# from app import routes, models
-
-class Item(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    title = db.Column(db.String(100), nullable=True)
-    price = db.Column(db.Integer, nullable=True)
-    isActive = db.Column(db.Boolean, default=True)
-    text = db.Column(db.Text, nullable=True)
-
-    def __repr__(self):
-        return '<Item {}>'.format(self.title)
-
-
-class User(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(64), index=True, unique=True)
-    email = db.Column(db.String(120), index=True, unique=True)
-    password_hash = db.Column(db.String(128))
-
-    def __repr__(self):
-        return '<User {}>'.format(self.username)
-
-
-@app.route('/')
-def index():
-    items = Item.query.order_by(Item.price).all()
-    return render_template('index.html', data=items)
-
-
-@app.route('/about')
-def about():
-    return render_template('about.html')
-
-
-@app.route('/buy/<int:id>')
-def buy(id):
-    item = Item.query.get(id)
-    api = Api(merchant_id=1396424,
-              secret_key='test')
-    checkout = Checkout(api=api)
-    data = {
-        "currency": "USD",
-        "amount": str(item.price) + '00'
-    }
-    url = checkout.url(data).get('checkout_url')
-    return redirect(url)
-
-
-@app.route('/create', methods=['POST', 'GET'])
-def create():
-    if request.method == 'POST':
-        title = request.form['title']
-        price = request.form['price']
-        text = request.form['text']
-        item = Item(title=title, price=price, text=text)
-        try:
-            db.session.add(item)
-            db.session.commit()
-            return redirect('/')
-        except:
-            return "Ошибка"
-    else:
-        return render_template('create.html')
-
-
-@app.route('/delete/<int:id>')
-def delete(id):
-    item = Item.query.get(id)
-    try:
-        db.session.delete(item)
-        db.session.commit()
-        return redirect('/')
-    except:
-        return "Ошибка"
-
-
-@app.route('/login', methods=['GET', 'POST'])
-def login():
-    form = LoginForm()
-    if form.validate_on_submit():
-        flash('Login requested for user {}, remember_me={}'.format(
-            form.username.data, form.remember_me.data))
-        return redirect(url_for('index'))
-    return render_template('login.html', title='Sign In', form=form)
-
-
-if __name__ == '__main__':
-    app.run(debug=True)
