+++ app/database.py
@@ -1,0 +1,6 @@
+from flask import app
+from flask_migrate import Migrate
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()
+migrate = Migrate(app, db)
