+++ requipments/base.txt
@@ -1,17 +1,23 @@
 alembic==1.4.2
 attrs==19.3.0
+automium==0.2.6
+automium-web==0.1.1
+blinker==1.4
 certifi==2020.6.20
 chardet==3.0.4
 click==7.1.2
 cloudipsp==1.0.2
 Flask==1.1.2
+Flask-DebugToolbar==0.11.0
 Flask-Migrate==2.5.3
 Flask-SQLAlchemy==2.4.4
 Flask-WTF==0.14.3
+gunicorn==20.0.4
 idna==2.10
 iniconfig==1.0.1
 itsdangerous==1.1.0
 Jinja2==2.11.2
+legacy==0.1.6
 Mako==1.1.3
 MarkupSafe==1.1.1
 migrate==0.3.8
