+++ requipments/development.txt
@@ -1,0 +1,2 @@
+-r base.txt
+flask-debugtoolbar
