+++ requipments/production.txt
@@ -1,0 +1,2 @@
+-r base.txt
+gunicorn
